requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-countries==5.5
django-crispy-forms==1.7.2
django-environ==0.4.5
idna==2.8
oauthlib==3.1.0
Pillow==6.2.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
stripe==2.37.2
urllib3==1.25.6
